feat: implement DepartmentSection and NewSemesterModal components; add SemesterCard for semester management; establish semester model for backend integration
